refactor: remove old duplicate detector
