+++ packages/dht/src/dht/DuplicateDetector.ts
@@ -1,52 +1,48 @@
-import { BloomFilter } from 'bloomfilter'
+import { Message } from "../proto/packages/dht/protos/DhtRpc"
 
-// Link for calculating false positive probabilities
-// https://www.di-mgt.com.au/bloom-calculator.html
+type QueueEntry = [timeStamp: number, value: string, senderId: string, message?: Message]
+
 export class DuplicateDetector {
-    currentFilter: BloomFilter
-    nextFilter?: BloomFilter
-    counter: number
-    private numOfHashFunctions: number
-    private numOfBits: number
-    private nextFilterFillingLimit: number
-    private resetLimit: number
 
-    // False positives at 0.05% at maximum capacity with default values
+    private values: Set<string> = new Set()
+    private queue: Array<QueueEntry> = []
+    private maxAge: number
+    private maxNumberOfValues: number
+
     constructor(
-        numOfBits = 2 ** 15,
-        numOfHashFunctions = 16,
-        nextFilterFillingLimit = 1050,
-        resetLimit = 2100
+        maxNumberOfValues: number,
+        maxAgeInSeconds: number
     ) {
-        this.currentFilter = new BloomFilter(numOfBits, numOfHashFunctions)
-        this.counter = 0
-        this.numOfHashFunctions = numOfHashFunctions
-        this.numOfBits = numOfBits
-        this.nextFilterFillingLimit = nextFilterFillingLimit
-        this.resetLimit = resetLimit
+        this.maxNumberOfValues = maxNumberOfValues
+        this.maxAge = maxAgeInSeconds * 1000
     }
 
-    add(value: string): void {
-        if (!this.nextFilter
-            && this.counter >= this.nextFilterFillingLimit
-            && this.counter < this.resetLimit
-        ) {
-            this.nextFilter = new BloomFilter(this.numOfBits, this.numOfHashFunctions)
-        } else if (this.counter >= this.resetLimit) {
-            this.counter = 0
-            this.currentFilter = this.nextFilter!
-            this.nextFilter = undefined
+    public add(value: string, senderId: string, message?: Message): void {
+        this.values.add(value)
+        if (message) {
+            this.queue.push([Date.now(), value, senderId, message])
+        } else {
+            this.queue.push([Date.now(), value, senderId])
         }
-        this.currentFilter.add(value)
-        if (this.nextFilter) {
-            this.nextFilter.add(value)
-        }
-        this.counter += 1
+        this.cleanUp()
     }
 
-    isMostLikelyDuplicate(value: string): boolean {
-        return this.currentFilter.test(value)
+    public isMostLikelyDuplicate(value: string): boolean {
+        return this.values.has(value)
+    }
+
+    private cleanUp(): void {
+        const currentTime = Date.now()
+
+        while (this.queue.length > 0 && (this.queue.length > this.maxNumberOfValues ||
+            (currentTime - this.queue[0][0]) > this.maxAge)) {
+            const oldestEntry = this.queue.shift()
+            this.values.delete(oldestEntry![1])
+        }
+    }
+
+    public size(): number {
+        return this.values.size
     }
 
 }
-
